pyproject.toml
[project]
name = "document_writer"
version = "0.1.0"
description = "Document Expansion Writer - AI-powered document research and writing tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "openai",
    "loguru",
    "typer",
    "python-dotenv",
    "tavily-python",
    "rich",
    "dirtyjson"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "ruff",
    "mypy"
]

[project.scripts]
docwriter = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.hatch.build.targets.wheel]
packages = ["src"]
